[ 투사체 궤적 구현 작업 ] [ 창 프리팹 수정 ] [ 창, 활 이미지 추가 : effect.png ] [ Misc.RandFloat 정상동작 안하는 문제 수정 : 정수를 실수로 형변환 시켜 계산함 ] [ 창객체에서 Start 함수가 호출 안되는 문제 수정 : 객체 비활성시 Start 메시지를 안보낸다. Start 함수 안에서 객체 비활성 처리하게 함 ] [ Shot 객체 : 그림자 크기 조정 처리 추가 , 소스 정리 ]

+++ Assets/Script/0_Utility/Misc.cs
@@ -589,12 +589,12 @@ namespace Utility
 
         //.NextDouble() 과 같은 함수이다.
         //returns a random double between zero and 1
-        static public double RandFloat()
+        static public float RandFloat()
         {
             //https://docs.microsoft.com/ko-kr/dotnet/api/system.random.next?view=netframework-4.7.2#System_Random_Next
             //MaxValue 는 rand.Next 가 나올수 있는 값보다 1 큰 값이다. 
             //1 큰값으로 나누기 때문에 1.0 에 결코 도달하지 못한다 
-            return (rand.Next() / (int.MaxValue));
+            return ((float)rand.Next() / (float)(int.MaxValue));
         }
 
         //RandFloat를 사용하기 때문에 최대값은 포함 되지 않는다.
